Split BBEditor and BBContainerView out into two pieces
