+++ tests/src/gui/MainWindowTest.cpp
@@ -1,0 +1,20 @@
+//
+// Created by Nicholas Wertzberger on 6/30/16.
+//
+
+#include "QTestSuite.h"
+#include "MainWindow.h"
+
+class MainWindowTest : QTestSuite
+{
+    Q_OBJECT
+private slots:
+
+    void KeyPressEventTest()
+    {
+        // TODO
+    }
+
+} MainWindowTests;
+
+#include "MainWindowTest.moc"
