+++ tests/src/gui/editors/BBEditorTest.cpp
@@ -1,0 +1,22 @@
+//
+// Created by Nicholas Wertzberger on 7/4/16.
+//
+
+#include "QTestSuite.h"
+#include "BBEditor.h"
+
+class BBEditorTest : QTestSuite
+{
+    Q_OBJECT
+private slots:
+
+    void AddTrack_AddsTracksInCorrectOrder()
+    {
+        //BBEditor * editor = new BBEditor((BBTrackContainer *) NULL);
+
+        //delete editor;
+    }
+
+} BBEditorTests;
+
+#include "BBEditorTest.moc"
